Added UserId with posts and Protected resources with Authorization

+++ backend/routes/posts.js
@@ -37,7 +37,8 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -70,11 +71,16 @@ multer({storage: storage}).single('image'),
     content: req.body.content,
     imagePath: imagePath
   });
-  console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result =>{
-    res.status(200).json({
-      message: 'Post updated succesfully'
-    }); // Everything is OK : 201
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result =>{
+    if (result.nModified > 0) {
+      res.status(200).json({
+        message: 'Post updated succesfully.'
+      }); // Everything is OK : 201
+    } else {
+      res.status(401).json({
+        message: 'Sorry! your post can not be updated due to non-authorization.'
+      }); // Everything is OK : 201
+    }
   });
 });
 
@@ -118,9 +124,17 @@ checkAuth, (req, res, next) =>{
   //   console.log(result);
   //   res.status(200).json({ message: 'Post deleted'});
   // });
-  Post.deleteOne({_id:req.params.id}).then(result =>{
+  Post.deleteOne({_id:req.params.id, creator: req.userData.userId}).then(result =>{
     console.log(result);
-    res.status(200).json({ message: 'Post deleted'});
+    if (result.n > 0) {
+      res.status(200).json({
+        message: 'Post deleted succesfully.'
+      }); // Everything is OK : 201
+    } else {
+      res.status(401).json({
+        message: 'Sorry! your post can not be deleted due to non-authorization.'
+      }); // Everything is OK : 201
+    }
   });
 });
